remove basestore.oldTransactableHandle
